Adds pocket view logic in it's own script
