+++ Assets/Scripts/SS3D/Systems/Storage/UI/InventoryUi.cs
@@ -34,7 +34,6 @@ namespace SS3D.Systems.Storage.UI
             Assert.IsNotNull(Inventory);
 
             HandsUi.Hands = Inventory.Hands;
-            SetUpPocketUI();
 
             Inventory.ContainerOpened += InventoryOnContainerOpened;
             Inventory.ContainerClosed += InventoryOnContainerClosed;
@@ -88,45 +87,6 @@ namespace SS3D.Systems.Storage.UI
             {
                 UiElement = uiElement;
                 Container = container;
-            }
-        }
-
-        private void SetUpPocketUI()
-        {
-            // Destroy existing elements and replace them by prefabs set up with their containers.
-            var childrenToDestroy = new List<GameObject>();
-            for (int i = 0; i < PocketParent.transform.childCount; i++)
-            {
-                if (PocketParent.transform.GetChild(i).gameObject.name.Contains("Pocket"))
-                {
-                    childrenToDestroy.Add(PocketParent.transform.GetChild(i).gameObject);
-                }
-            }
-
-            foreach (var child in childrenToDestroy)
-            {
-                DestroyImmediate(child);
-            }
-
-            var InventoryContainers = Inventory.gameObject.GetComponentsInChildren<ContainerDescriptor>();
-            var PocketContainers = new List<ContainerDescriptor>();
-            foreach (var container in InventoryContainers)
-            {
-                if (container.ContainerType is ContainerType.Pocket)
-                    PocketContainers.Add(container);
-            }
-            if (PocketContainers.Count == 0)
-            {
-                throw new ApplicationException("no container of type pocket is present on " +
-                    "the inventory's game object or any of it's children.");
-            }
-            foreach (var container in PocketContainers)
-            {
-                var attachedContainer = container.AttachedContainer;
-                GameObject handElement = Instantiate(PocketPrefab, PocketParent, false);
-                SingleItemContainerSlot slot = handElement.GetComponent<SingleItemContainerSlot>();
-                slot.Inventory = Inventory;
-                slot.Container = attachedContainer;
             }
         }
     }
