+++ Assets/Scripts/SS3D/Systems/Storage/UI/PocketView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d16d9c9e1fac9c84da0a1db31c0c4294
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
